Add JSP support.
As Undertow does not directly support JSPs, it is needed to remove it, go back to Tomcat and add the jasper library.
In order to avoid problems when the war is deployed to Widfly, Tomcat is still excluded but added when run as a Boot app
as it is referenced as provided. This will allow run the code as a Spring Boot app inside Tomcat and as a war inside Wildfly
without clashing.